# src/docs/swagger.yaml
openapi: 3.0.3
info:
  title: Yum-Gott Platform API
  description: 'API for the Yum-Gott platform, supporting customer and restaurant owner authentication, profile management, product management, and category management.'
  version: 1.0.0
servers:
  - url: 'http://localhost:3000'
    description: Local development server
  - url: 'https://production-server-yumgott-30534781840.europe-west1.run.app'
    description: Production server
  - url: 'https://yum-gott.onrender.com'
    description: Development server
paths:
...
  /api/products:
    post:
      tags:
        - Products
      summary: Create a new product
      description: Creates a new product for a restaurant owner.
      operationId: createProduct
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                categoryName:
                  type: string
                  description: 'Name of the category. If the category does not exist, it will be created automatically.'
                  example: Main Course
                productName:
                  type: string
                  description: Name of the product
                  example: Margherita Pizza
                description:
                  type: string
                  description: Product description
                  example: Classic pizza with tomato sauce and mozzarella
                price:
                  type: number
                  description: Base price of the product
                  example: 100
                discount:
                  type: number
                  description: Discount percentage
                  example: 10
                sizeOptions:
                  type: object
                  description: 'Optional array of size options. In multipart/form-data, must be a JSON string (e.g., ''[{"name":"Medium","additionalPrice":50},{"name":"Large","additionalPrice":70}]'').'
                  example: '[{"name":"Medium","additionalPrice":50},{"name":"Large","additionalPrice":70}]'
                image:
                  type: string
                  format: binary
                  description: Product image file (JPEG or PNG)
                options:
                  type: object
                  description: |
                    Optional array of product options (as JSON string). Each option can include name, required, and values (array of {name, additionalPrice}). Do NOT send id when creating new options/values; it will be generated automatically.
                  example: '[{"name":"Toppings","required":true,"values":[{"name":"Extra Cheese","additionalPrice":2.5}]}]'
              required:
                - categoryName
                - productName
                - description
                - price
                - image
      responses:
        '201':
          description: Product created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Product created successfully
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: Invalid input
...
  /api/categories:
    post:
      tags:
        - Categories
      summary: Create a new category
      description: Allows restaurant owners to create a new category for their products.
      operationId: createCategory
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                  maxLength: 100
                  example: Main Course
      responses:
        '201':
          description: Category created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Category created successfully
                  data:
                    type: object
                    properties:
                      id:
                        type: string
                        format: uuid
                        example: 550e8400-e29b-41d4-a716-446655440000
                      name:
                        type: string
                        example: Main Course
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
...
  /api/opening-hours:
    post:
      tags:
        - Opening Hours
      summary: Create opening hours for a restaurant
      description: 'Allows restaurant owners to define one or more Working_hours per day (e.g., morning and evening shifts) in a single request.'
      operationId: createOpeningHours
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - day
                - isClosed
                - Working_hours
              properties:
                day:
                  type: string
                  enum:
                    - Monday
                    - Tuesday
                    - Wednesday
                    - Thursday
                    - Friday
                    - Saturday
                    - Sunday
                  example: Monday
                Working_hours:
                  type: array
                  items:
                    type: object
                    properties:
                      startTime:
                        type: string
                        format: time
                        example: '08:00:00'
                      endTime:
                        type: string
                        format: time
                        example: '12:00:00'
                  example:
                    - startTime: '08:00:00'
                      endTime: '12:00:00'
                    - startTime: '14:00:00'
                      endTime: '18:00:00'
                isClosed:
                  type: boolean
                  example: false
            examples:
              multipleWorkingHours:
                summary: Multiple working hours in one request
                value:
                  day: Monday
                  Working_hours:
                    - startTime: '08:00:00'
                      endTime: '12:00:00'
                    - startTime: '14:00:00'
                      endTime: '18:00:00'
                  isClosed: false
      responses:
        '201':
          description: Opening hours created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Opening hours created successfully
                  data:
                    type: object
                    properties:
                      id:
                        type: string
                        format: uuid
                      restaurantOwnerId:
                        type: string
                        format: uuid
                      day:
                        type: string
                        example: Monday
                      Working_hours:
                        type: array
                        items:
                          type: object
                          properties:
                            startTime:
                              type: string
                              format: time
                              example: '08:00:00'
                            endTime:
                              type: string
                              format: time
                              example: '12:00:00'
                        example:
                          - startTime: '08:00:00'
                            endTime: '12:00:00'
                          - startTime: '14:00:00'
                            endTime: '18:00:00'
                      isClosed:
                        type: boolean
                        example: false
                      createdAt:
                        type: string
                        format: date-time
                      updatedAt:
                        type: string
                        format: date-time
        '400':
          description: Bad request
        '403':
          description: Forbidden
...
  /api/videos:
    post:
      summary: Create a new video
      tags:
        - Videos
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                publicId:
                  type: string
                  description: Public ID of the video
                  example: yum_gott/videos/yum_gott_1752456506674
                secureUrl:
                  type: string
                  format: uri
                  description: Secure URL of the video
                  example: 'https://res.cloudinary.com/dbipfyr2o/video/upload/v1752456506674/yum_gott/videos/yum_gott_1752456506674.mp4'
                restaurantName:
                  type: string
                  description: Name of the restaurant
                  example: Falafel Ala Kefk
                phoneNumber:
                  type: string
                  description: Phone number associated with the video
                  example: '2054353'
                network:
                  type: string
                  enum:
                    - MTN
                    - Syriatel
                  description: Network provider
                  example: MTN
                invoiceImage:
                  type: string
                  format: binary
                  description: Invoice image file (JPEG or PNG)
                  example: invoice.jpg
              required:
                - publicId
                - secureUrl
                - restaurantName
                - phoneNumber
                - network
                - invoiceImage
      responses:
        '201':
          description: Video created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Video created successfully
                  data:
                    type: object
                    properties:
                      id:
                        type: string
                        format: uuid
                      userId:
                        type: string
                        format: uuid
                      publicId:
                        type: string
                      secureUrl:
                        type: string
                        format: uri
                      restaurantName:
                        type: string
                      phoneNumber:
                        type: string
                      network:
                        type: string
                        enum:
                          - MTN
                          - Syriatel
                      invoiceImage:
                        type: string
                        format: uri
                      statusVideo:
                        type: string
                        enum:
                          - pending
                          - rejected
                          - accepted
                      likesCount:
                        type: integer
                        description: Number of likes on the video
                        example: 42
                      createdAt:
                        type: string
                        format: date-time
                      updatedAt:
                        type: string
                        format: date-time
                    required:
                      - id
                      - userId
                      - publicId
                      - secureUrl
                      - restaurantName
                      - phoneNumber
                      - network
                      - invoiceImage
                      - statusVideo
                      - likesCount
                      - createdAt
                      - updatedAt
        '400':
          description: Bad request (validation error)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized (invalid or missing token)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden (only customers can create videos)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/comments:
    post:
      tags:
        - Comments
      summary: Create a new comment on a video
      description: Creates a new comment on a video. Requires authentication.
      operationId: createComment
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                videoId:
                  type: string
                  format: uuid
                  description: ID of the video to comment on
                  example: 123e4567-e89b-12d3-a456-426614174000
                content:
                  type: string
                  description: Comment content
                  minLength: 1
                  maxLength: 1000
                  example: Great video! The food looks amazing.
              required:
                - videoId
                - content
      responses:
        '201':
          description: Comment created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Comment created successfully
                  data:
                    $ref: '#/components/schemas/Comment'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Video not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/payGate:
    get:
      tags:
        - PayGate
      summary: List active pay gates
      description: Returns all active payment gateways as an array of strings.
      operationId: listPayGates
      responses:
        '200':
          description: Active pay gates retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Active pay gates
                  data:
                    type: array
                    items:
                      type: string
                      example: PAYMENT_X
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/orders:
    post:
      summary: Create a new order
      tags:
        - Orders
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                customerId:
                  type: string
                product:
                  type: object
                  properties:
                    product:
                      type: object
                      properties:
                        categoryName:
                          type: string
                        productName:
                          type: string
                        description:
                          type: string
                        price:
                          type: number
                        discount:
                          type: number
                        sizeOptions:
                          type: array
                          items:
                            type: object
                            properties:
                              name:
                                type: string
                              additionalPrice:
                                type: number
                        restaurantOwnerId:
                          type: string
                        createdAt:
                          type: string
                          format: date-time
                        updatedAt:
                          type: string
                          format: date-time
                    options:
                      type: array
                      items:
                        type: object
                        properties:
                          option:
                            type: object
                            properties:
                              name:
                                type: string
                              required:
                                type: boolean
                              createdAt:
                                type: string
                                format: date-time
                              updatedAt:
                                type: string
                                format: date-time
                          values:
                            type: array
                            items:
                              type: object
                              properties:
                                name:
                                  type: string
                                additionalPrice:
                                  type: number
                                createdAt:
                                  type: string
                                  format: date-time
                                updatedAt:
                                  type: string
                                  format: date-time
                orderDate:
                  type: string
                  format: date-time
                status:
                  type: string
                  enum:
                    - pending
                    - paid
                    - shipped
                    - completed
                    - cancelled
                createdAt:
                  type: string
                  format: date-time
                updatedAt:
                  type: string
                  format: date-time
      responses:
        '201':
          description: Order created
          content:
            application/json:
              schema:
                $ref: '#/paths/~1api~1orders/post/requestBody/content/application~1json/schema'
    get:
      summary: Get all orders for the authenticated customer
      tags:
        - Orders
      security:
        - cookieAuth: []
      responses:
        '200':
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/paths/~1api~1orders/post/requestBody/content/application~1json/schema'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: accessToken
  schemas:
...
    Comment:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 123e4567-e89b-12d3-a456-426614174000
        content:
          type: string
          example: Great video! The food looks delicious.
        videoId:
          type: string
          format: uuid
          example: 123e4567-e89b-12d3-a456-426614174001
        userId:
          type: string
          format: uuid
          example: 123e4567-e89b-12d3-a456-426614174002
        userType:
          type: string
          enum:
            - customer
            - restaurant_owner
          example: customer
        userName:
          type: string
          description: Display name of the commenter (customer name or restaurant name)
          example: John Doe
        createdAt:
          type: string
          format: date-time
          example: '2025-01-01T00:00:00Z'
        updatedAt:
          type: string
          format: date-time
          example: '2025-01-01T00:00:00Z'
      required:
        - id
        - content
        - videoId
        - userId
        - userType
        - createdAt
        - updatedAt
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: An error occurred
      required:
        - success
        - message
...
  responses:
    BadRequestError:
      description: Bad request
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              message:
                type: string
                example: Bad request
    UnauthorizedError:
      description: Unauthorized
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              message:
                type: string
                example: Unauthorized
    ForbiddenError:
      description: Forbidden
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              message:
                type: string
                example: Forbidden
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              message:
                type: string
                example: Internal server error